Updated code to new structure
After merge with develop there is a new app form_data so I embedded all new
files in this app.
