+++ api/views.py
@@ -1,8 +1,8 @@
 # encoding: utf-8
-from django.views import generic
 from rest_framework import viewsets
 
-from form_data.models import FormData
+from form_data import models
+
 from .serializers import FormDataSerializer
 
 
@@ -20,12 +20,3 @@ class FormDataViewSet(viewsets.ModelViewSet):
         serializer_class = self.get_serializer_class()
         kwargs['context'] = self.get_serializer_context()
         return serializer_class(*args, **kwargs)
-
-
-class FormDataListView(generic.ListView):
-    """View that shows all from of the current user."""
-    model = models.FormData
-    ordering = 'title'
-
-    def get_queryset(self):
-        return models.FormData.objects.user_related(self.request.user.id)
