+++ form_data/urls.py
@@ -5,5 +5,6 @@ from . import views
 
 urlpatterns = patterns('',
     url(r'create/$', views.FormDataCreateView.as_view(), name='form_data_create'),
-    url(r'update/(?P<pk>\d+)/$', views.FormDataUpdateView.as_view(), name='form_data_update')
+    url(r'update/(?P<pk>\d+)/$', views.FormDataUpdateView.as_view(), name='form_data_update'),
+    url(r'^list/$', views.FormDataListView.as_view(), name='formdata_list'),
 )
