+++ hopper/config/urls.py
@@ -4,7 +4,7 @@ from django.conf.urls.static import static
 from django.contrib import admin
 from rest_framework import routers
 
-from api.views import FormDataListView, FormDataViewSet
+from api.views import FormDataViewSet
 
 router = routers.DefaultRouter()
 router.register(r'v1', FormDataViewSet, base_name='api')
@@ -13,7 +13,6 @@ urlpatterns = patterns('',
     url(r'^grappelli/', include('grappelli.urls')),
     url(r'^admin/', include(admin.site.urls)),
     url(r'^auth/$', include('rest_framework.urls', namespace='rest_framework')),
-    url(r'^list/$', FormDataListView.as_view(), name='formdata_list'),
     url(r'^api/', include(router.urls)),
     url(r'^forms/', include('form_data.urls')),
     url(r'^', include('landingpage.urls'))
