+++ tests/api_tests/test_forms.py
@@ -1,7 +1,7 @@
 # encoding: utf-8
 import pytest
 
-from api.forms import HopperForm
+from form_data.forms import HopperForm
 
 
 @pytest.mark.django_db
