+++ tests/form_data_tests/test_form_data_views.py
@@ -10,10 +10,8 @@ def test_form_data_create_login_required(client):
 
 
 @pytest.mark.django_db
-def test_form_data_create_login_post(credentials, user, client):
-    username, password = credentials
+def test_form_data_create_login_post(user, login, client):
     from form_data.models import FormData
-    assert client.login(username=username, password=password)
     assert not FormData.objects.count()
     response = client.post(reverse('form_data_create'), data={'title': 'Test'})
     assert response.status_code == 302
@@ -21,19 +19,15 @@ def test_form_data_create_login_post(credentials, user, client):
 
 
 @pytest.mark.django_db
-def test_form_data_update_ownership_required(form_with_other_user, credentials, user, client):
-    username, password = credentials
-    assert client.login(username=username, password=password)
+def test_form_data_update_ownership_required(form_with_other_user, user, login, client):
     assert form_with_other_user.author != user
     response = client.get(reverse('form_data_update', kwargs={'pk': form_with_other_user.pk}))
     assert response.status_code == 403
 
 
 @pytest.mark.django_db
-def test_form_data_update_ownership_create_token(model, credentials, user, client):
+def test_form_data_update_ownership_create_token(model, user, login, client):
     from rest_framework.authtoken.models import Token
-    username, password = credentials
-    assert client.login(username=username, password=password)
     assert model.author == user
     assert not Token.objects.count()
     client.get(reverse('form_data_update', kwargs={'pk': model.pk}))
